Add AspectMock for mocking built-in functions
It's possible to do this without the library, but the library makes it
simpler, so we don't have to have a bunch of extra files in the test code
base for each of the mocked methods in the namespace, and then manually
require them. This is much cleaner and I think makes it more obvious what
the mock is doing.