Use specific versions of require-dev dependencies
It may help to prevent issues such as
https://github.com/ramsey/uuid/commit/bd7f7f3135b1cd751b8be06b66a721e13f3d5b06
when a BC break is introduced in a non-semver manner.

On one hand, it means that from time to time, versions need to be
updated manually. But on the other hand it allows to handle updates of
dev dependencies intentionally, without colliding with other changes.